requirements.txt
typing-extensions>=4.11,<5
pyyaml>=6,<7
openstacksdk>=3,<4
